requirements.txt
ansible==9.0.1
ansible-compat==4.1.10
ansible-core==2.16.0
ansible-lint==6.22.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.0.1
attrs==23.1.0
bcrypt==4.1.1
black==23.11.0
bracex==2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-help-colors==0.9.4
cryptography==41.0.7
deprecated==1.2.14
dill==0.3.7
distro==1.8.0
docker==6.1.3
enrich==1.2.7
filelock==3.13.1
flake8==6.1.0
hvac==2.0.0
idna==3.6
isort==5.12.0
jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
markupsafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.0
molecule==6.0.2
molecule-plugins[docker]==23.5.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
psychicoctowinner @ git+https://github.com/arpanrec/psychicoctowinner.git
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pyflakes==3.1.0
pygithub==2.1.1
pygments==2.17.2
pyjwt[crypto]==2.8.0
pylint==3.0.2
pynacl==1.5.0
pyopenssl==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==4.2.0
python-gnupg==0.5.1
pyyaml==6.0.1
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.7.0
rpds-py==0.13.2
ruamel-yaml==0.18.5
ruamel-yaml-clib==0.2.8
selinux==0.3.0
six==1.16.0
subprocess-tee==0.4.1
tomlkit==0.12.3
typing-extensions==4.8.0
urllib3==2.1.0
wcmatch==8.5
websocket-client==1.6.4
wrapt==1.16.0
yamllint==1.33.0
